benchmarks/embedding_collection/torchrec_dlrm/requirements.txt
fbgemm-gpu==0.4.1
torchmetrics==1.1.0
torchrec==0.4.0